# openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "InviteResponse": {
        "properties": {
          "allow_guests": {
            "title": "Allow Guests",
            "type": "boolean"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "email_bound": {
            "title": "Email Bound",
            "type": "boolean"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "plan",
              "group"
            ],
            "title": "Kind",
            "type": "string"
          },
          "max_uses": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Uses"
          },
          "purpose": {
            "enum": [
              "join",
              "claim"
            ],
            "title": "Purpose",
            "type": "string"
          },
          "requires_approval": {
            "title": "Requires Approval",
            "type": "boolean"
          },
          "role": {
            "title": "Role",
            "type": "string"
          },
          "state": {
            "enum": [
              "active",
              "revoked"
            ],
            "title": "State",
            "type": "string"
          },
          "target_participant_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Participant Id"
          },
          "use_count": {
            "title": "Use Count",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "kind",
          "purpose",
          "role",
          "allow_guests",
          "requires_approval",
          "email_bound",
          "target_participant_id",
          "max_uses",
          "use_count",
          "expires_at",
          "state",
          "created_at"
        ],
        "title": "InviteResponse",
        "type": "object"
      },
      "InviteTokenRequest": {
        "additionalProperties": false,
        "properties": {
          "token": {
            "maxLength": 128,
            "minLength": 20,
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "token"
        ],
        "title": "InviteTokenRequest",
        "type": "object"
      },
...
      "PublicProblemResponse": {
        "description": "Typed-client-safe problem fields guaranteed across every API error.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instance"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "title",
          "status",
          "code"
        ],
        "title": "PublicProblemResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "API for plans with your people.",
    "title": "Beluno API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/groups/{group_id}/invites": {
      "get": {
        "operationId": "list_group_invites_v1_groups__group_id__invites_get",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Group Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/InviteResponse"
                  },
                  "title": "Response List Group Invites V1 Groups  Group Id  Invites Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "Bearer authentication is missing, invalid, expired, or revoked."
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "The current relationship does not allow this action."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "The resource is absent or not visible to the caller."
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "The request conflicts with the current resource state."
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "The request does not satisfy the contract."
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "Too many requests; retry after the Retry-After interval."
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "A dependency or feature is temporarily unavailable."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Group Invites",
        "tags": [
          "invites"
        ]
      },
      "post": {
        "operationId": "create_group_invite_v1_groups__group_id__invites_post",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Group Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GroupInviteCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedInviteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "Bearer authentication is missing, invalid, expired, or revoked."
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "The current relationship does not allow this action."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "The resource is absent or not visible to the caller."
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "The request conflicts with the current resource state."
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "The request does not satisfy the contract."
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "Too many requests; retry after the Retry-After interval."
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "A dependency or feature is temporarily unavailable."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Group Invite",
        "tags": [
          "invites"
        ]
      }
    },
    "/v1/groups/{group_id}/invites/{invite_id}": {
      "delete": {
        "operationId": "revoke_group_invite_v1_groups__group_id__invites__invite_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Group Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "invite_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Invite Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InviteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "Bearer authentication is missing, invalid, expired, or revoked."
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "The current relationship does not allow this action."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "The resource is absent or not visible to the caller."
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "The request conflicts with the current resource state."
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "The request does not satisfy the contract."
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "Too many requests; retry after the Retry-After interval."
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "A dependency or feature is temporarily unavailable."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Revoke Group Invite",
        "tags": [
          "invites"
        ]
      }
    },
...
    "/v1/invites/preview": {
      "post": {
        "operationId": "preview_invite_v1_invites_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitePreviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "Bearer authentication is missing, invalid, expired, or revoked."
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "The current relationship does not allow this action."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "The resource is absent or not visible to the caller."
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "The request conflicts with the current resource state."
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "The request does not satisfy the contract."
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "Too many requests; retry after the Retry-After interval."
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "A dependency or feature is temporarily unavailable."
          }
        },
        "summary": "Preview Invite",
        "tags": [
          "invites"
        ]
      }
    },
    "/v1/invites/redeem": {
      "post": {
        "description": "Join a plan or group. Without an account, plan invites that allow guests\ncreate a guest session (returned in ``session``).",
        "operationId": "redeem_invite_v1_invites_redeem_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RedeemInviteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RedeemInviteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "Bearer authentication is missing, invalid, expired, or revoked."
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "The current relationship does not allow this action."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "The resource is absent or not visible to the caller."
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "The request conflicts with the current resource state."
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "The request does not satisfy the contract."
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "Too many requests; retry after the Retry-After interval."
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "A dependency or feature is temporarily unavailable."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Redeem Invite",
        "tags": [
          "invites"
        ]
      }
    },
...
    "/v1/plans/{plan_id}/invites/{invite_id}": {
      "delete": {
        "operationId": "revoke_plan_invite_v1_plans__plan_id__invites__invite_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "plan_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Plan Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "invite_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Invite Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InviteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "Bearer authentication is missing, invalid, expired, or revoked."
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "The current relationship does not allow this action."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "The resource is absent or not visible to the caller."
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "The request conflicts with the current resource state."
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "The request does not satisfy the contract."
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "Too many requests; retry after the Retry-After interval."
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "A dependency or feature is temporarily unavailable."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Revoke Plan Invite",
        "tags": [
          "invites"
        ]
      }
    },
...
  }
}